move resync functions in networkchain module and explicit rename
